重构配置文件 将配置文件分为主配置文件 流体列表与物品列表
