+++ src/main/java/com/EyeOfHarmonyBuffer/Mixins/EyeOfHarmonyEU.java
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import com.EyeOfHarmonyBuffer.Config;
+import com.EyeOfHarmonyBuffer.Config.Config;
 
 import gregtech.api.recipe.check.CheckRecipeResult;
 import tectech.recipe.EyeOfHarmonyRecipe;
