+++ src/main/java/com/EyeOfHarmonyBuffer/Mixins/EyeOfHarmonyItem.java
@@ -9,7 +9,8 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import com.EyeOfHarmonyBuffer.Config;
+import com.EyeOfHarmonyBuffer.Config.ItemConfig;
+import com.EyeOfHarmonyBuffer.Config.MainConfig;
 import com.EyeOfHarmonyBuffer.OutputProcessing.CustomItemStackLong;
 import com.EyeOfHarmonyBuffer.utils.ItemInfo;
 
@@ -24,21 +25,21 @@ public abstract class EyeOfHarmonyItem {
 
         try {
             // 检查是否启用额外产出
-            if (!Config.EOHItemInPut) {
+            if (!MainConfig.EOHItemInPut) {
                 System.out.println("额外产出已禁用");
                 return;
             }
 
             // 检查配置的输出物品列表是否为空
-            if (Config.outputItems == null || Config.outputItems.isEmpty()) {
+            if (ItemConfig.outputItems == null || ItemConfig.outputItems.isEmpty()) {
                 System.err.println("错误：Config.outputItems 为空。");
                 return;
             }
 
-            System.out.println("要输出的物品数量：" + Config.outputItems.size());
+            System.out.println("要输出的物品数量：" + ItemConfig.outputItems.size());
 
             // 遍历配置中的每个物品
-            for (ItemInfo itemInfo : Config.outputItems) {
+            for (ItemInfo itemInfo : ItemConfig.outputItems) {
                 try {
                     // 处理物品，并使用 CustomItemStackLong 包装物品
                     CustomItemStackLong customItemStack = new CustomItemStackLong(itemInfo);
