+++ src/main/java/com/EyeOfHarmonyBuffer/Mixins/EyeOfHarmonyMixin.java
@@ -5,7 +5,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import com.EyeOfHarmonyBuffer.Config;
+import com.EyeOfHarmonyBuffer.Config.MainConfig;
 
 import tectech.thing.metaTileEntity.multi.MTEEyeOfHarmony;
 
@@ -15,6 +15,6 @@ public class EyeOfHarmonyMixin {
     @Inject(method = "recipeProcessTimeCalculator", at = @At("HEAD"), cancellable = true)
     private void recipeProcessTimeCalculator(long recipeTime, long recipeSpacetimeCasingRequired,
         CallbackInfoReturnable<Integer> cir) {
-        cir.setReturnValue(Config.EOHtime);
+        cir.setReturnValue(MainConfig.EOHtime);
     }
 }
